+++ contracts/fx-portal/token/ERC20/FxERC20MintBurn.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: MIT
-pragma solidity 0.8.19;
+pragma solidity 0.8.28;
 
 import {IForwarderRegistry} from "@animoca/ethereum-contracts/contracts/metatx/interfaces/IForwarderRegistry.sol";
 import {IFxERC20MintBurn} from "./interfaces/IFxERC20MintBurn.sol";