+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 2.0.0
+
+### Breaking changes
+
+- Upgraded to `@openzeppelin/contracts@5.2.0`.
+- Upgraded to `solc@0.8.28`.
+- Upgraded to `@animoca/ethereum-contracts@4.0.0`.
+
+### Improvements
+
+- Updated to latest dependencies.
+
 ## 1.0.0
 
 ### Breaking changes